dbeaver/pro#7820 Move function settings to another class
